requirements.txt
warcio==1.7.4
beautifulsoup4==4.12.3
lxml==5.2.1
langid==1.1.6
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
pandas>=2.2
pyarrow>=16.0
python-dateutil>=2.9
numpy==1.26.4
matplotlib==3.8.4
statsmodels==0.14.2
